[feat] 유저 생성 & 수정 & 본인 조회

+++ application/balance-eat-api/src/main/kotlin/org/balanceeat/api/diet/DietV1Payload.kt
@@ -4,7 +4,6 @@ import io.swagger.v3.oas.annotations.media.Schema
 import jakarta.validation.constraints.NotNull
 import jakarta.validation.constraints.Positive
 import org.balanceeat.domain.diet.Diet
-import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.LocalDateTime
 
@@ -29,7 +28,7 @@ class DietV1Request {
         @field:NotNull(message = "섭취량은 필수입니다")
         @field:Positive(message = "섭취량은 양수여야 합니다")
         @Schema(description = "섭취량", requiredMode = Schema.RequiredMode.REQUIRED)
-        val servingSize: BigDecimal,
+        val servingSize: Double,
         
         @field:NotNull(message = "섭취량 단위는 필수입니다")
         @Schema(description = "섭취량 단위", requiredMode = Schema.RequiredMode.REQUIRED)
@@ -38,25 +37,25 @@ class DietV1Request {
         @field:NotNull(message = "칼로리는 필수입니다")
         @field:Positive(message = "칼로리는 양수여야 합니다")
         @Schema(description = "칼로리", requiredMode = Schema.RequiredMode.REQUIRED)
-        val calories: BigDecimal,
+        val calories: Double,
         
         @Schema(description = "탄수화물 (g)", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-        val carbohydrates: BigDecimal? = null,
+        val carbohydrates: Double? = null,
         
         @Schema(description = "단백질 (g)", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-        val protein: BigDecimal? = null,
+        val protein: Double? = null,
         
         @Schema(description = "지방 (g)", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-        val fat: BigDecimal? = null,
+        val fat: Double? = null,
         
         @Schema(description = "당류 (g)", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-        val sugar: BigDecimal? = null,
+        val sugar: Double? = null,
         
         @Schema(description = "나트륨 (mg)", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-        val sodium: BigDecimal? = null,
+        val sodium: Double? = null,
         
         @Schema(description = "식이섬유 (g)", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-        val fiber: BigDecimal? = null,
+        val fiber: Double? = null,
         
         @field:NotNull(message = "식사 날짜는 필수입니다")
         @Schema(description = "식사 날짜", requiredMode = Schema.RequiredMode.REQUIRED)
@@ -76,15 +75,15 @@ class DietV1Response {
         val mealTypeName: String,
         val foodName: String,
         val foodBrand: String?,
-        val servingSize: BigDecimal,
+        val servingSize: Double,
         val servingUnit: String,
-        val calories: BigDecimal,
-        val carbohydrates: BigDecimal?,
-        val protein: BigDecimal?,
-        val fat: BigDecimal?,
-        val sugar: BigDecimal?,
-        val sodium: BigDecimal?,
-        val fiber: BigDecimal?,
+        val calories: Double,
+        val carbohydrates: Double?,
+        val protein: Double?,
+        val fat: Double?,
+        val sugar: Double?,
+        val sodium: Double?,
+        val fiber: Double?,
         val mealDate: LocalDate,
         val consumedAt: LocalDateTime?,
         val createdAt: LocalDateTime
@@ -92,10 +91,10 @@ class DietV1Response {
     
     data class Summary(
         val mealDate: LocalDate,
-        val totalCalories: BigDecimal,
-        val totalCarbohydrates: BigDecimal?,
-        val totalProtein: BigDecimal?,
-        val totalFat: BigDecimal?,
+        val totalCalories: Double,
+        val totalCarbohydrates: Double?,
+        val totalProtein: Double?,
+        val totalFat: Double?,
         val mealCounts: Map<Diet.MealType, Int>,
         val diets: List<Info>
     )
