+++ domain/src/main/kotlin/org/balanceeat/domain/user/UserCommand.kt
@@ -1,6 +1,5 @@
 package org.balanceeat.domain.user
 
-import java.math.BigDecimal
 
 class UserCommand {
 
@@ -9,16 +8,16 @@ class UserCommand {
         val name: String,
         val gender: User.Gender,
         val age: Int,
-        val height: BigDecimal,
-        val weight: BigDecimal,
+        val height: Double,
+        val weight: Double,
         val email: String? = null,
         val activityLevel: User.ActivityLevel? = null,
-        val smi: BigDecimal? = null,
-        val fatPercentage: BigDecimal? = null,
-        val targetWeight: BigDecimal? = null,
+        val smi: Double? = null,
+        val fatPercentage: Double? = null,
+        val targetWeight: Double? = null,
         val targetCalorie: Int? = null,
-        val targetSmi: BigDecimal? = null,
-        val targetFatPercentage: BigDecimal? = null,
+        val targetSmi: Double? = null,
+        val targetFatPercentage: Double? = null,
         val providerId: String? = null,
         val providerType: String? = null
     )
@@ -28,14 +27,14 @@ class UserCommand {
         val email: String? = null,
         val gender: User.Gender? = null,
         val age: Int? = null,
-        val height: BigDecimal? = null,
-        val weight: BigDecimal? = null,
+        val height: Double? = null,
+        val weight: Double? = null,
         val activityLevel: User.ActivityLevel? = null,
-        val smi: BigDecimal? = null,
-        val fatPercentage: BigDecimal? = null,
-        val targetWeight: BigDecimal? = null,
+        val smi: Double? = null,
+        val fatPercentage: Double? = null,
+        val targetWeight: Double? = null,
         val targetCalorie: Int? = null,
-        val targetSmi: BigDecimal? = null,
-        val targetFatPercentage: BigDecimal? = null
+        val targetSmi: Double? = null,
+        val targetFatPercentage: Double? = null
     )
 }
